Feat: Combine the settings and personalization page
